requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
tuspyserver>=1.0
aiosqlite>=0.20
Pillow>=10.0
pillow-heif>=0.16
python-multipart>=0.0.9
